+++ processor/src/test/fixtures/expected/Arez_BasicObservableModel.java
@@ -12,7 +12,7 @@ import javax.annotation.Nonnull;
 import org.realityforge.braincheck.Guards;
 
 @Generated("arez.processor.ArezProcessor")
-public final class Arez_BasicObservableModel extends BasicObservableModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
+final class Arez_BasicObservableModel extends BasicObservableModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
   private static volatile int $$arezi$$_nextId;
 
   private final ComponentKernel $$arezi$$_kernel;