+++ processor/src/test/fixtures/expected/Arez_DisposingModel.java
@@ -13,7 +13,7 @@ import org.realityforge.braincheck.Guards;
 
 @Generated("arez.processor.ArezProcessor")
 @SuppressWarnings("unchecked")
-public final class Arez_DisposingModel extends DisposingModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
+final class Arez_DisposingModel extends DisposingModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
   private static volatile int $$arezi$$_nextId;
 
   private final ComponentKernel $$arezi$$_kernel;