+++ processor/src/test/fixtures/expected/Arez_OverrideNamesInModel.java
@@ -17,7 +17,7 @@ import org.realityforge.braincheck.Guards;
 
 @Generated("arez.processor.ArezProcessor")
 @SuppressWarnings("unchecked")
-public final class Arez_OverrideNamesInModel extends OverrideNamesInModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
+final class Arez_OverrideNamesInModel extends OverrideNamesInModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
   private static volatile int $$arezi$$_nextId;
 
   private final ComponentKernel $$arezi$$_kernel;