+++ processor/src/test/fixtures/expected/com/example/computable_value_ref/Arez_MultiComputableValueRefModel.java
@@ -15,7 +15,7 @@ import org.realityforge.braincheck.Guards;
 
 @Generated("arez.processor.ArezProcessor")
 @SuppressWarnings("unchecked")
-public final class Arez_MultiComputableValueRefModel extends MultiComputableValueRefModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
+final class Arez_MultiComputableValueRefModel extends MultiComputableValueRefModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
   private static volatile int $$arezi$$_nextId;
 
   private final ComponentKernel $$arezi$$_kernel;