+++ processor/src/test/fixtures/expected/com/example/inject/Arez_DefaultCtorModel.java
@@ -15,7 +15,7 @@ import javax.inject.Inject;
 import org.realityforge.braincheck.Guards;
 
 @Generated("arez.processor.ArezProcessor")
-public final class Arez_DefaultCtorModel extends DefaultCtorModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
+final class Arez_DefaultCtorModel extends DefaultCtorModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
   private static volatile int $$arezi$$_nextId;
 
   private final ComponentKernel $$arezi$$_kernel;