+++ processor/src/test/fixtures/expected/com/example/pre_dispose/Arez_MultiPreDisposeModel.java
@@ -12,7 +12,7 @@ import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 
 @Generated("arez.processor.ArezProcessor")
-public final class Arez_MultiPreDisposeModel extends MultiPreDisposeModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
+final class Arez_MultiPreDisposeModel extends MultiPreDisposeModel implements Disposable, Identifiable<Integer>, DisposeNotifier {
   private static volatile int $$arezi$$_nextId;
 
   private final ComponentKernel $$arezi$$_kernel;